# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messaging service API documentation",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Current version"
        }
    ],
    "paths": {
...
        "/notification": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Display list of notifications",
                "operationId": "1d827a59b0def54b0d6fec288dcae119",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Provide a unique ID for specified client to filter messages",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of notifications",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Store a newly created notification in storage.",
                "operationId": "5e996908224e5e9c706d3fd8ae035edc",
                "requestBody": {
                    "description": "Notification object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Notification"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notification/bulk": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Notification bulk store",
                "description": "Store a multiple newly created notifications in storage.",
                "operationId": "a21b22f0b964a6c2c103a16b292fece6",
                "requestBody": {
                    "description": "Messages to store",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notification/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Display the specified resource.",
                "description": "Returns single notification",
                "operationId": "9219c0fe0f3e08e9cc79237d6d9fb769",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID specific to this notification",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found"
                    },
                    "500": {
                        "description": "Server error. Most likely id is not in propper UUID format. See error message for details."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Notification": {
                "description": "Notification class",
                "required": [
                    "clientId",
                    "channel",
                    "message"
                ],
                "properties": {
                    "id": {
                        "description": "primary key ID for the message",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": "true"
                    },
                    "clientId": {
                        "description": "Id of the client",
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "description": "Message delivery channel",
                        "type": "string",
                        "enum": [
                            "sms",
                            "email"
                        ],
                        "example": "sms"
                    },
                    "message": {
                        "description": "Mewssage for client",
                        "type": "text",
                        "example": "Molestiae veniam quia ut. Dolores ea quod in. Nam et aut corrupti dolorum. Id et et odit beatae laborum nesciunt."
                    },
                    "status": {
                        "description": "Delivery status of the message: pending, processing, sent, failed",
                        "type": "text",
                        "readOnly": "true",
                        "example": "sent"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}